+++ terkin/datalogger.py
@@ -4,6 +4,7 @@
 # License: GNU General Public License, Version 3
 import time
 import machine
+
 from terkin import __version__
 from terkin import logging
 from terkin.configuration import TerkinConfiguration
@@ -12,7 +13,6 @@ from terkin.network import SystemWiFiMetrics
 from terkin.sensor import SensorManager, AbstractSensor
 from terkin.sensor.system import SystemMemoryFree, SystemTemperature, SystemBatteryLevel, SystemUptime
 from terkin.util import dformat, gc_disabled, ddformat
-import sys
 
 log = logging.getLogger(__name__)
 
@@ -117,8 +117,7 @@ class TerkinDatalogger:
         self.device.watchdog.reconfigure_minimum_timeout(15000)
         if not self.settings.get('main.fastboot', False):
             self.device.power_off_lte_modem()
-        if sys.platform in ['WiPy', 'LoPy', 'GPy', 'FiPy']:
-            self.device.power_off_bluetooth()
+        self.device.power_off_bluetooth()
         self.device.watchdog.resume()
 
         log.info('Starting %s', self.application_info.fullname)
@@ -288,7 +287,7 @@ class TerkinDatalogger:
 
         system_sensors = [
             SystemMemoryFree,
-            #SystemTemperature,
+            SystemTemperature,
             SystemBatteryLevel,
             SystemUptime,
         ]