+++ terkin/logging.py
@@ -2,14 +2,17 @@
 # (c) 2019 Richard Pobering <richard@hiveeyes.org>
 # (c) 2019 Andreas Motl <andreas@hiveeyes.org>
 # License: GNU General Public License, Version 3
-import sys, utime
+import sys
+import utime
 import logging
 from logging import Logger, StreamHandler, Formatter, _level, _loggers
 from logging import DEBUG, INFO, WARNING, ERROR, CRITICAL
+from mboot import MicroPythonPlatform
+from terkin.util import GenericChronometer, PycomChronometer, get_platform_info
 
 # Keep track of time since boot.
-from terkin.util import GenericChronometer, PycomChronometer
-if sys.platform in ['WiPy', 'LoPy', 'GPy', 'FiPy']:
+platform_info = get_platform_info()
+if platform_info.vendor == MicroPythonPlatform.Pycom:
     _chrono = PycomChronometer()
 else:
     _chrono = GenericChronometer()
@@ -18,13 +21,11 @@ else:
 class TimedLogRecord(logging.LogRecord):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        if sys.platform in ['WiPy', 'LoPy', 'GPy', 'FiPy']:
-            try:
-                self.tdelta = _chrono.read()
-            except:
-                self.tdelta = None
-        else:
-            self.tdelta = utime.time()
+        try:
+            self.tdelta = _chrono.read()
+        except:
+            self.tdelta = 0
+
 
 class ExtendedLogger(Logger):