+++ terkin/sensor/system.py
@@ -2,13 +2,18 @@
 # (c) 2019 Richard Pobering <richard@hiveeyes.org>
 # (c) 2019 Andreas Motl <andreas@hiveeyes.org>
 # License: GNU General Public License, Version 3
+import sys
 import time
 from machine import ADC, enable_irq, disable_irq
 from micropython import const
+
+from mboot import MicroPythonPlatform
 from terkin import logging
-import sys
+from terkin.util import get_platform_info
 
 log = logging.getLogger(__name__)
+platform_info = get_platform_info()
+
 
 # Todo: Make this configurable.
 #log.setLevel(logging.DEBUG)
@@ -142,7 +147,7 @@ class SystemBatteryLevel:
         log.debug('Reading battery level on pin {} with voltage divider {}/{}'.format(self.pin, self.resistor_r1, self.resistor_r2))
 
         # read samples
-        if sys.platform in ['WiPy', 'LoPy', 'GPy', 'FiPy']:
+        if platform_info.vendor == MicroPythonPlatform.Pycom:
             self.adc.init()
             adc_channel = self.adc.channel(attn=ADC.ATTN_6DB, pin=self.pin)
             irq_state = disable_irq()
@@ -166,10 +171,11 @@ class SystemBatteryLevel:
             adc_variance += (sample - adc_mean) ** 2
         adc_variance /= (self.adc_sample_count - 1)
 
-        if sys.platform in ['WiPy', 'LoPy', 'GPy', 'FiPy']:
+        if platform_info.vendor == MicroPythonPlatform.Pycom:
             raw_voltage = adc_channel.value_to_voltage(4095)
             mean_voltage = adc_channel.value_to_voltage(int(adc_mean))
-        else:   # TODO: make this work for esp32
+        else:
+            # FIXME: Make this work for vanilla ESP32.
             raw_voltage = 0.0
             mean_voltage = 0.0
         mean_variance = (adc_variance * 10 ** 6) // (adc_mean ** 2)
@@ -181,14 +187,15 @@ class SystemBatteryLevel:
         log.debug("SystemBatteryLevel: 10**6*Variance/(Mean**2) of ADC readings = %15.13f" % mean_variance)
 
         resistor_sum = self.resistor_r1 + self.resistor_r2
-        if sys.platform in ['WiPy', 'LoPy', 'GPy', 'FiPy']:
+        if platform_info.vendor == MicroPythonPlatform.Pycom:
             voltage_millivolt = (adc_channel.value_to_voltage(int(adc_mean))) * resistor_sum / self.resistor_r2
-        else:   # TODO: make this work for esp32
+        else:
+            # FIXME: Make this work for vanilla ESP32.
             voltage_millivolt = 0.0            
         voltage_volt = voltage_millivolt / 1000.0
 
         # Shut down ADC channel.
-        if sys.platform in ['WiPy', 'LoPy', 'GPy', 'FiPy']:
+        if platform_info.vendor == MicroPythonPlatform.Pycom:
             adc_channel.deinit()
 
         log.debug('Battery level: {}'.format(voltage_volt))